update restoration sites
