+++ src/components/ui/popup/index.tsx
@@ -2,6 +2,10 @@ import { ReactNode } from 'react';
 
 import { Popup as PopupReactMapGL } from 'react-map-gl';
 
+import Icon from 'components/ui/icon';
+
+import CLOSE_SVG from 'svgs/ui/close.svg?sprite';
+
 const Popup = ({
   children,
   popUpPosition,
@@ -33,9 +37,18 @@ const Popup = ({
       latitude={latitude || null}
       onClose={onClose}
       closeOnClick={false}
-      className="c-restoration-popup"
+      className="relative min-w-fit p-0"
     >
-      {children}
+      <div className="relative flex h-full  min-w-fit grow flex-col justify-between divide-y divide-gray-100 lg:max-w-sm">
+        {children}
+        <button
+          type="button"
+          onClick={onClose}
+          className="absolute top-4 -right-10 z-50 h-11 w-10 cursor-pointer items-center justify-end rounded-r-[20px] bg-white/70 backdrop-blur-sm focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2"
+        >
+          <Icon icon={CLOSE_SVG} className="ml-1 h-6 w-6" description="Close" />
+        </button>
+      </div>
     </PopupReactMapGL>
   );
 };
