+++ src/components/ui/select/index.tsx
@@ -42,12 +42,7 @@ const SelectContent = React.forwardRef<
       position={position}
       {...props}
     >
-      <SelectPrimitive.Viewport
-        className={cn(
-          'p-1',
-          position === 'popper' && 'w-full min-w-[var(--radix-select-trigger-width)]'
-        )}
-      >
+      <SelectPrimitive.Viewport className={cn('p-1', position === 'popper' && 'w-full')}>
         {children}
       </SelectPrimitive.Viewport>
     </SelectPrimitive.Content>
