+++ src/containers/datasets/restoration-sites/map-popup.tsx
@@ -1,0 +1,44 @@
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from 'components/ui/collapsible';
+import { RestorationSitesPopUp } from 'types/map';
+
+const PopupRestorationSites = ({ info }: { info: RestorationSitesPopUp }) => (
+  <Collapsible className="min-w-[375px]">
+    <CollapsibleTrigger>
+      <h2>RESTORATION SITES</h2>
+    </CollapsibleTrigger>
+    <CollapsibleContent>
+      <div className="flex w-full flex-col space-y-2 border-none p-6 font-sans shadow-none">
+        {info.cluster && (
+          <p className="text-sm text-black/85">
+            <span>
+              There are <strong>{info.point_count}</strong> restoration sites in this location.{' '}
+            </span>
+            <span className="font-extralight">Zoom in to view more details about each site</span>
+          </p>
+        )}
+
+        <div className="flex w-full justify-between font-sans">
+          {info.site_name && (
+            <div className="flex flex-col">
+              <span className="text-left text-sm font-semibold text-brand-800">Site</span>
+              <span className="text-left text-xxs font-light uppercase leading-5 text-black/85">
+                {info.site_name}
+              </span>
+            </div>
+          )}
+
+          {info.landscape_name && (
+            <div className="flex flex-col">
+              <span className="text-left text-sm font-semibold text-brand-800">Landscape</span>
+              <span className="text-left text-xxs font-light uppercase leading-5 text-black/85">
+                {info.landscape_name}
+              </span>
+            </div>
+          )}
+        </div>
+      </div>
+    </CollapsibleContent>
+  </Collapsible>
+);
+
+export default PopupRestorationSites;
