--- src/containers/map/restoration-popup/index.tsx
@@ -1,59 +1,0 @@
-import { useCallback, useState } from 'react';
-
-import cn from 'lib/classnames';
-
-import Details from 'containers/map/restoration-popup/sections/details';
-import EcosystemServices from 'containers/map/restoration-popup/sections/ecosystem-services';
-import RestorationScores from 'containers/map/restoration-popup/sections/restoration-scores';
-
-import type { RestorationPopUp } from 'types/map';
-
-const PopupRestoration = ({
-  restorationPopUpInfo,
-  className,
-}: {
-  restorationPopUpInfo: {
-    popupInfo: RestorationPopUp;
-  };
-  className?: string;
-}) => {
-  const [open, setOpen] = useState('');
-
-  const handleClick = useCallback(
-    (id: string) => {
-      if (!open || id !== open) {
-        setOpen(id);
-      } else {
-        setOpen(null);
-      }
-    },
-    [open]
-  );
-
-  return (
-    <div
-      className={cn({
-        'c-restoration-popup relative flex w-[500px] flex-col items-start bg-white': true,
-        [className]: !!className,
-      })}
-    >
-      <RestorationScores
-        data={restorationPopUpInfo?.popupInfo}
-        isOpen={open === 'restoration'}
-        handleClick={() => handleClick('restoration')}
-      />
-      <Details
-        data={restorationPopUpInfo?.popupInfo}
-        isOpen={open === 'details'}
-        handleClick={() => handleClick('details')}
-      />
-      <EcosystemServices
-        data={restorationPopUpInfo?.popupInfo}
-        isOpen={open === 'ecosystem'}
-        handleClick={() => handleClick('ecosystem')}
-      />
-    </div>
-  );
-};
-
-export default PopupRestoration;
