+++ src/containers/navigation/menu/index.tsx
@@ -125,8 +125,8 @@ const Menu = () => {
               <ContactForm className="text-left text-2lg font-light hover:text-brand-800" />
 
               <Collapsible open={isOpen} onOpenChange={setIsOpen} className="space-y-4">
-                <CollapsibleTrigger>
-                  <div className="flex items-center space-x-4 text-2lg font-light hover:text-brand-800">
+                <CollapsibleTrigger showExpandIcon={false} className="p-0">
+                  <div className="flex w-full items-center space-x-4 text-2lg font-light hover:text-brand-800">
                     <span className={cn({ 'text-brand-800': isOpen })}>Resources</span>
                     <HiChevronDown
                       className={cn({
@@ -137,18 +137,20 @@ const Menu = () => {
                     <span className="sr-only">Toggle</span>
                   </div>
                 </CollapsibleTrigger>
-                <CollapsibleContent className="flex flex-col space-y-2 border-l border-l-grey-400/20 px-6">
-                  {RESOURCES_LINKS.map(({ label, href }) => (
-                    <a
-                      key={label}
-                      href={href}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className="text-2lg font-light text-black/85 hover:text-brand-800"
-                    >
-                      {label}
-                    </a>
-                  ))}
+                <CollapsibleContent>
+                  <div className="flex flex-col space-y-2 border-l border-l-grey-400/20 px-6">
+                    {RESOURCES_LINKS.map(({ label, href }) => (
+                      <a
+                        key={label}
+                        href={href}
+                        target="_blank"
+                        rel="noopener noreferrer"
+                        className="text-2lg font-light text-black/85 hover:text-brand-800"
+                      >
+                        {label}
+                      </a>
+                    ))}
+                  </div>
                 </CollapsibleContent>
               </Collapsible>
             </div>
@@ -187,7 +189,7 @@ const Menu = () => {
             </motion.div>
           )}
         </AnimatePresence>
-        <DialogClose className="" />
+        <DialogClose />
       </DialogContent>
     </Dialog>
   );
