+++ src/layouts/desktop/index.tsx
@@ -17,7 +17,7 @@ import { useRecoilValue } from 'recoil';
 import { useLocation } from 'containers/datasets/locations/hooks';
 import type { LocationTypes } from 'containers/datasets/locations/types';
 import MapContainer from 'containers/map';
-import WelcomeIntroMessage from 'containers/welcome-message';
+// import WelcomeIntroMessage from 'containers/welcome-message';
 import WidgetsContainer from 'containers/widgets';
 
 const DesktopLayout = () => {
@@ -83,7 +83,7 @@ const DesktopLayout = () => {
         <MapContainer mapId={`default-desktop-${isPrintingId}`} />
         <WidgetsContainer />
 
-        <WelcomeIntroMessage />
+        {/* <WelcomeIntroMessage /> */}
       </div>
     </div>
   );
